+++ src/routes/gamerules/components/index.css
@@ -1,7 +1,3 @@
-.rules{
-    margin-bottom: 17px;
-}
-
 .slider {
     display: flex;
     justify-content: space-around;
